@@ -2,24 +2,30 @@
   "name": "@studiometa/js-toolkit-demo",
   "version": "3.0.0-alpha.3",
   "private": true,
-  "type": "commonjs",
+  "type": "module",
   "scripts": {
     "dev": "meta dev",
     "watch": "meta watch",
-    "build": "meta build"
+    "build": "meta build",
+    "lint": "npm run lint:eslint && npm run lint:stylelint && npm run lint:prettier",
+    "lint:eslint": "eslint --cache ./src/js",
+    "lint:stytelint": "stylelint --cache ./src/**/*.{scss,vue}",
+    "lint:prettier": "prettier --cache ./src/**/*.twig"
   },
   "dependencies": {
-    "@studiometa/eslint-config": "^3.1.3",
-    "@studiometa/stylelint-config": "^3.0.2",
+    "@studiometa/eslint-config": "^4.0.1",
+    "@studiometa/prettier-config": "^4.0.0",
+    "@studiometa/stylelint-config": "^4.0.0",
     "@studiometa/tailwind-config": "^2.1.0",
     "@studiometa/ui": "^0.2.47",
-    "@studiometa/webpack-config": "^5.3.0",
-    "@studiometa/webpack-config-preset-prototyping": "^5.3.0",
-    "eslint": "^8.57.0",
+    "@studiometa/webpack-config": "^6.0.1",
+    "@studiometa/webpack-config-preset-prototyping": "^6.0.1",
+    "eslint": "^9.0.0",
     "postcss": "^8.4.38",
-    "prettier": "^2.8.8",
+    "prettier": "^3.2.5",
     "stylefire": "^7.0.3",
-    "stylelint": "^15.11.0",
+    "stylelint": "^16.3.1",
     "tailwindcss": "^3.4.3"
-  }
+  },
+  "prettier": "@studiometa/prettier-config"
 }
