@@ -1,4 +1,6 @@
-module.exports = {
-  presets: [require('@studiometa/tailwind-config')],
+import { config } from '@studiometa/tailwind-config';
+
+export default {
+  presets: [config],
   content: ['src/templates/**/*.twig', 'src/**/*.js'],
 };
